Move wait_for_tasks helper func to capsule_mixin
