+++ robottelo/hosts.py
@@ -1134,7 +1134,11 @@ class ContentHost(Host, ContentHostMixins):
             raise ContentHostError(
                 f'Error during registration, command output: {cmd_result.stdout}'
             )
-        self.subscription_manager_attach_pool([settings.subscription.rhn_poolid])[0]
+        cmd_result = self.subscription_manager_attach_pool([settings.subscription.rhn_poolid])[0]
+        if cmd_result.status != 0:
+            raise ContentHostError(
+                f'Error during pool attachment, command output: {cmd_result.stdout}'
+            )
 
 
 class Capsule(ContentHost, CapsuleMixins):
@@ -1543,36 +1547,15 @@ class Satellite(Capsule, SatelliteMixins):
             )
         self.add_rex_key(self)
 
-    def wait_for_tasks(
-        self, search_query, search_rate=1, max_tries=10, poll_rate=None, poll_timeout=None
-    ):
-        """Search for tasks by specified search query and poll them to ensure that
-        task has finished.
-
-        :param search_query: Search query that will be passed to API call.
-        :param search_rate: Delay between searches.
-        :param max_tries: How many times search should be executed.
-        :param poll_rate: Delay between the end of one task check-up and
-            the start of the next check-up. Parameter for ``sat.api.ForemanTask.poll()`` method.
-        :param poll_timeout: Maximum number of seconds to wait until timing out.
-            Parameter for ``sat.api.ForemanTask.poll()`` method.
-        :return: List of ``sat.api.ForemanTasks`` entities.
-        :raises: ``AssertionError``. If not tasks were found until timeout.
-        """
-        for _ in range(max_tries):
-            tasks = self.api.ForemanTask().search(query={'search': search_query})
-            if len(tasks) > 0:
-                for task in tasks:
-                    task.poll(poll_rate=poll_rate, timeout=poll_timeout)
-                break
-            else:
-                time.sleep(search_rate)
-        else:
-            raise AssertionError(f"No task was found using query '{search_query}'")
-        return tasks
-
     def register_host_custom_repo(self, module_org, rhel_contenthost, repo_urls):
-        """Register content host to Satellite and sync repos"""
+        """Register content host to Satellite and sync repos
+
+        :param module_org: Org where contenthost will be registered.
+        :param rhel_contenthost: contenthost to be register with Satellite.
+        :param repo_urls: List of URLs to be synced and made available to contenthost
+            via subscription-manager.
+        :return: None
+        """
         # Create a new product, sync appropriate client and other passed repos on satellite
         rhelver = rhel_contenthost.os_version.major
         prod = self.api.Product(
